First push after significant modifications; fixing error codes, begining to refactor importing data, removing massive data file, etc
